+++ pages/friends.tsx
@@ -1,15 +1,14 @@
-import { Typography } from '@mui/material';
 import { NextPage } from 'next';
 import React from 'react';
 import Top from '../components/Top';
-import InteractiveList from './FriendsList';
+import InteractiveList from '../components/friends/FriendsList';
 
 const Friend: NextPage = () => {
     return (
         <div>
             <Top page={2} />
             <div>
-                <InteractiveList></InteractiveList> 
+                <InteractiveList></InteractiveList>
             </div>
         </div>
     )
